backend/requirements.txt
fastapi[standard]>=0.115,<1.0
pydantic-settings>=2.0,<3.0
sqlalchemy>=2.0,<3.0
asyncpg>=0.29,<1.0
aiosqlite>=0.20,<1.0
greenlet>=3.0,<4.0
pytest>=8.0,<9.0
pytest-asyncio>=0.23,<1.0
pytest-cov>=5.0,<7.0
flake8>=7.0,<8.0
httpx>=0.27,<1.0
bcrypt>=4.0,<5.0
PyJWT>=2.8,<3.0
slowapi>=0.1.9,<1.0
openai>=1.0,<2.0
